backend/pyproject.toml
[project]
name = "oci-llm-comparison-backend"
version = "1.0.0"
description = "Backend API for OCI LLM Comparison Demo"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "langchain-oci==0.1.0",
    "langchain-core>=0.3.15,<0.4",
    "python-dotenv==1.0.0",
    "pydantic>=2.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

